+++ src/walls.c
@@ -6,7 +6,7 @@
 /*   By: jrocha-v <jrocha-v@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/21 17:23:26 by jrocha-v          #+#    #+#             */
-/*   Updated: 2024/05/22 15:59:19 by jrocha-v         ###   ########.fr       */
+/*   Updated: 2024/05/29 12:59:31 by jrocha-v         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,12 @@
 void	render_walls(t_data *data)
 {
 	int i;
-	int	dist_pp;
+	double	dist_pp;
 	t_point line_start;
 	t_point	line_end;
 	
 	i = -1;
-	dist_pp = (WIN_W / 2) / tan(FOV_ANG / 2);
+	dist_pp = (WIN_W / 2.0) / tan(FOV_ANG / 2.0);
 	while (++i < NUM_RAYS)
 	{
 		data->rays[i].distance = data->rays[i].distance * cos(data->rays[i].angle - data->player.rot_ang);
